Feat: ThemeZ Fix: group token
You can now change the theme of MyChat.

Themes can be created using ThemeZ lua.
If you have a MyThemeZ.lua in your config folder we will load those. otherwise we use the included ones.

Fixed a bug with using GROUP token and not being in a group. 

GROUP token will search for the Group members names as Standard or all lowercase.