feat:
Added a default channel Consider

This is a catch all to hold consider messages.
these messages will be color coded based on Con Color of your target.

I think i got all of the consider messages added, let me know if I am missing anything.

This channel will auto-magically be added to your existing configs if missing.  So feel free to accidently delete it. =P

The only way to edit this is to have the channel tab enabled and right click to get to the events config page.  or through the config file itself. Colors on this don't matter as we overwrite the settings with con color.