fix: LinkDB empty raw data table
if the link db has an empty raw data table from not using /link /update it was causing issues with not having a name to pull from.

There is merit to building a personal DB from scratch especially on EMU. as this weeds out items not on your server.

As such we not will strip the item name out of the item link.